# pkg/metastore: Reduce allocs in MakeLocationID