Reduced allocations in `MakeLocationID`.

```
$ go test -run=^$ -bench ^BenchmarkMakeLocationID$ -benchmem ./pkg/metastore -count=1
...
# new
BenchmarkMakeLocationID-12    	 3224959	       370.4 ns/op	      96 B/op	       1 allocs/op
# old
BenchmarkMakeLocationID-12    	 1859330	       647.9 ns/op	     560 B/op	      10 allocs/op
```

This helped to shave 17K allocations per operation in `NormalizeWriteRawRequest`.

```
$ benchstat bench-old.txt bench-new.txt
name                         old time/op    new time/op    delta
NormalizeWriteRawRequest-12    43.9ms ± 7%    43.1ms ± 5%  -1.89%  (p=0.001 n=47+46)

name                         old alloc/op   new alloc/op   delta
NormalizeWriteRawRequest-12    53.5MB ± 1%    52.7MB ± 1%  -1.62%  (p=0.000 n=50+50)

name                         old allocs/op  new allocs/op  delta
NormalizeWriteRawRequest-12      412k ± 0%      395k ± 0%  -4.12%  (p=0.000 n=50+50)
```

Together with https://github.com/parca-dev/parca/pull/3061 it should save ~31K/operation.