@kakkoyun 2023-05-16
@marselester Besides the unit tests and benchmarks, could you confirm everything works as expected with a smoke test? (actually, running Parca with the agent and ensuring everything works visually) Sorry for the toil, we don't have a fully-fledged integration test for these parts.